fix indentation
beautiful jekyll uses by default the kramdown parser for markdown. It's not exactly identical to the github-flavoured markdown. You can change the markdown flavour in the config file: https://github.com/UBC-MDS/UBC-MDS.github.io/blob/master/_config.yml#L76

According to [this issue](https://github.com/gettalong/kramdown/issues/123), the correct way to have an indented code block inside a list is to indent the code block with 8 spaces (4 spaces for being inside a list, and 4 spaces for being a code block). The rest of the text inside that list needs to be indented by 4 spaces (not 2 spaces as was done before).  I haven't tested it, but this should fix it hopefully