German translation improvements
You did a amazing Job with this template. I really like it as base for my own project.
So I'd have one error and one request for improvement. All regarding the German translation in ui-text.yml.

[...]
comment_form_email_label : "E-Mail-Addresse"
[...]
There i one d too much. the word is 'E-Mail-Adresse.

The 'Sie' is very formal. Meanwhile even in Germany is the 'you' well accepted. Especially online it is common to use the more casual 'Du'. But as mentioned before is is not an error, it is just my idea to improve the theme.
The Change is what 'I've changed for my site.