Add UBC's required statement
The last bullet point in item 7 on UBC's ["Content and Distribution of Course Syllabi"](http://www.calendar.ubc.ca/vancouver/index.cfm?tree=3,328,0,0) requires this paragraph to appear on all course syllabi. Figured here would be the best place for it, since we point to this page from our syllabi.