Add margin-bottom to social media links
The social media links at the bottom of the page lack a margin-bottom
meaning that when there are many of them and they spill onto a second
line, there is no space between the rows, and this looks ugly. This
change adds a suitable margin and also reduces the footer's overall
margin-bottom to match the increase in item margin-bottom.